Move files within the gem
