+++ lib/darthjee/core_ext.rb
@@ -1,10 +1,10 @@
-require 'enumerable'
 require 'active_support/all'
-require 'hash'
-require 'symbol'
-require 'array'
-require 'numeric'
-require 'date_ext'
+require 'darthjee/core_ext/enumerable'
+require 'darthjee/core_ext/hash'
+require 'darthjee/core_ext/symbol'
+require 'darthjee/core_ext/array'
+require 'darthjee/core_ext/numeric'
+require 'darthjee/core_ext/date'
 
 module Darthjee
   module CoreExt
