+++ lib/darthjee/core_ext/array/hash_builder.rb
