+++ lib/darthjee/core_ext/enumerable.rb
