+++ lib/darthjee/core_ext/hash/key_changer.rb
