+++ lib/darthjee/core_ext/hash/value_changer.rb
