+++ lib/darthjee/core_ext/numeric.rb
